# apps/web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessReasonIn": {
        "properties": {
          "reason_code": {
            "enum": [
              "covering_doctor",
              "lab_review",
              "second_opinion",
              "administrative",
              "other"
            ],
            "title": "Reason Code",
            "type": "string"
          },
          "reason_text": {
            "maxLength": 300,
            "minLength": 5,
            "title": "Reason Text",
            "type": "string"
          }
        },
        "required": [
          "reason_code",
          "reason_text"
        ],
        "title": "AccessReasonIn",
        "type": "object"
      },
      "AdminChart": {
        "properties": {
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "view": {
            "const": "admin",
            "default": "admin",
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "patient"
        ],
        "title": "AdminChart",
        "type": "object"
      },
...
      "AllergyOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "reaction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reaction"
          },
          "severity": {
            "$ref": "#/components/schemas/AllergySeverity"
          },
          "substance": {
            "title": "Substance",
            "type": "string"
          }
        },
        "required": [
          "id",
          "substance",
          "reaction",
          "severity",
          "is_active"
        ],
        "title": "AllergyOut",
        "type": "object"
      },
      "AllergySeverity": {
        "enum": [
          "mild",
          "moderate",
          "severe"
        ],
        "title": "AllergySeverity",
        "type": "string"
      },
...
      "BreakGlassIn": {
        "properties": {
          "reason_code": {
            "enum": [
              "emergency",
              "unconscious",
              "severe_allergy",
              "system_outage",
              "other"
            ],
            "title": "Reason Code",
            "type": "string"
          },
          "reason_text": {
            "maxLength": 300,
            "minLength": 5,
            "title": "Reason Text",
            "type": "string"
          }
        },
        "required": [
          "reason_code",
          "reason_text"
        ],
        "title": "BreakGlassIn",
        "type": "object"
      },
      "BreakGlassItem": {
        "properties": {
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient": {
            "$ref": "#/components/schemas/PatientSummary"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "reason_text": {
            "title": "Reason Text",
            "type": "string"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "user_name": {
            "title": "User Name",
            "type": "string"
          },
          "user_role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "at",
          "user_id",
          "user_name",
          "user_role",
          "patient",
          "reason_code",
          "reason_text",
          "reviewed_at",
          "reviewed_by",
          "outcome"
        ],
        "title": "BreakGlassItem",
        "type": "object"
      },
      "BreakGlassReviewIn": {
        "properties": {
          "note": {
            "anyOf": [
              {
                "maxLength": 500,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "outcome": {
            "$ref": "#/components/schemas/ReviewOutcome"
          }
        },
        "required": [
          "outcome"
        ],
        "title": "BreakGlassReviewIn",
        "type": "object"
      },
...
      "ConditionOut": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "onset": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onset"
          },
          "status": {
            "$ref": "#/components/schemas/ConditionStatus"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "text",
          "onset",
          "status"
        ],
        "title": "ConditionOut",
        "type": "object"
      },
      "ConditionStatus": {
        "enum": [
          "active",
          "resolved",
          "inactive"
        ],
        "title": "ConditionStatus",
        "type": "string"
      },
...
      "DoctorChart": {
        "properties": {
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyOut"
            },
            "title": "Allergies",
            "type": "array"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "title": "Appointments",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionOut"
            },
            "title": "Conditions",
            "type": "array"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Documents",
            "type": "array"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation"
          },
          "lab_orders": {
            "items": {
              "$ref": "#/components/schemas/LabOrderOut"
            },
            "title": "Lab Orders",
            "type": "array"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "title": "Notes",
            "type": "array"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "prescriptions": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionOut"
            },
            "title": "Prescriptions",
            "type": "array"
          },
          "view": {
            "const": "doctor",
            "default": "doctor",
            "title": "View",
            "type": "string"
          },
          "vitals": {
            "items": {
              "$ref": "#/components/schemas/VitalOut"
            },
            "title": "Vitals",
            "type": "array"
          }
        },
        "required": [
          "patient",
          "appointments",
          "vitals",
          "allergies",
          "conditions",
          "notes",
          "prescriptions",
          "lab_orders",
          "documents"
        ],
        "title": "DoctorChart",
        "type": "object"
      },
      "DocumentOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "sha256",
          "created_at"
        ],
        "title": "DocumentOut",
        "type": "object"
      },
...
      "EmergencyChart": {
        "properties": {
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyOut"
            },
            "title": "Allergies",
            "type": "array"
          },
          "break_glass_id": {
            "format": "uuid",
            "title": "Break Glass Id",
            "type": "string"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionOut"
            },
            "title": "Conditions",
            "type": "array"
          },
          "current_medicines": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Current Medicines",
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "last_vitals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VitalOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "patient": {
            "$ref": "#/components/schemas/PatientSummary"
          },
          "view": {
            "const": "emergency",
            "default": "emergency",
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "break_glass_id",
          "expires_at",
          "patient",
          "allergies",
          "conditions",
          "current_medicines",
          "last_vitals"
        ],
        "title": "EmergencyChart",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LabChart": {
        "properties": {
          "patient": {
            "$ref": "#/components/schemas/PatientSummary"
          },
          "view": {
            "const": "lab",
            "default": "lab",
            "title": "View",
            "type": "string"
          },
          "worklist": {
            "items": {
              "$ref": "#/components/schemas/LabOrderOut"
            },
            "title": "Worklist",
            "type": "array"
          }
        },
        "required": [
          "patient",
          "worklist"
        ],
        "title": "LabChart",
        "type": "object"
      },
...
      "NurseChart": {
        "properties": {
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyOut"
            },
            "title": "Allergies",
            "type": "array"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientSummary"
          },
          "today_appointment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppointmentOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "view": {
            "const": "nurse",
            "default": "nurse",
            "title": "View",
            "type": "string"
          },
          "vitals": {
            "items": {
              "$ref": "#/components/schemas/VitalOut"
            },
            "title": "Vitals",
            "type": "array"
          }
        },
        "required": [
          "patient",
          "today_appointment",
          "vitals",
          "allergies"
        ],
        "title": "NurseChart",
        "type": "object"
      },
...
      "PatientChart": {
        "description": "What a patient sees of their own record: signed notes and prescriptions, released results.",
        "properties": {
          "allergies": {
            "items": {
              "$ref": "#/components/schemas/AllergyOut"
            },
            "title": "Allergies",
            "type": "array"
          },
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "title": "Appointments",
            "type": "array"
          },
          "conditions": {
            "items": {
              "$ref": "#/components/schemas/ConditionOut"
            },
            "title": "Conditions",
            "type": "array"
          },
          "consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentOut"
            },
            "title": "Documents",
            "type": "array"
          },
          "invoices": {
            "items": {
              "$ref": "#/components/schemas/InvoiceSummary"
            },
            "title": "Invoices",
            "type": "array"
          },
          "lab_orders": {
            "items": {
              "$ref": "#/components/schemas/LabOrderOut"
            },
            "title": "Lab Orders",
            "type": "array"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteOut"
            },
            "title": "Notes",
            "type": "array"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "prescriptions": {
            "items": {
              "$ref": "#/components/schemas/PrescriptionOut"
            },
            "title": "Prescriptions",
            "type": "array"
          },
          "view": {
            "const": "patient",
            "default": "patient",
            "title": "View",
            "type": "string"
          },
          "vitals": {
            "items": {
              "$ref": "#/components/schemas/VitalOut"
            },
            "title": "Vitals",
            "type": "array"
          }
        },
        "required": [
          "patient",
          "appointments",
          "vitals",
          "allergies",
          "conditions",
          "notes",
          "prescriptions",
          "lab_orders",
          "documents",
          "invoices",
          "consent"
        ],
        "title": "PatientChart",
        "type": "object"
      },
...
      "PatientSummary": {
        "properties": {
          "dob": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dob"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "mrn": {
            "title": "Mrn",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sex": {
            "$ref": "#/components/schemas/Sex"
          }
        },
        "required": [
          "id",
          "mrn",
          "name",
          "dob",
          "sex"
        ],
        "title": "PatientSummary",
        "type": "object"
      },
...
      "ReceptionChart": {
        "properties": {
          "appointments": {
            "items": {
              "$ref": "#/components/schemas/AppointmentOut"
            },
            "title": "Appointments",
            "type": "array"
          },
          "consent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConsentStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "invoices": {
            "items": {
              "$ref": "#/components/schemas/InvoiceSummary"
            },
            "title": "Invoices",
            "type": "array"
          },
          "patient": {
            "$ref": "#/components/schemas/PatientOut"
          },
          "view": {
            "const": "reception",
            "default": "reception",
            "title": "View",
            "type": "string"
          }
        },
        "required": [
          "patient",
          "appointments",
          "invoices",
          "consent"
        ],
        "title": "ReceptionChart",
        "type": "object"
      },
...
      "ReviewOutcome": {
        "enum": [
          "benign",
          "misuse",
          "unsure"
        ],
        "title": "ReviewOutcome",
        "type": "string"
      },
      "Role": {
        "enum": [
          "platform_admin",
          "clinic_admin",
          "reception",
          "nurse",
          "doctor",
          "lab_tech",
          "patient"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "Sex": {
        "enum": [
          "female",
          "male",
          "other",
          "unknown"
        ],
        "title": "Sex",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VitalOut": {
        "properties": {
          "bp_dia": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bp Dia"
          },
          "bp_sys": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bp Sys"
          },
          "height_cm": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "pulse": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pulse"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "rr": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rr"
          },
          "spo2": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spo2"
          },
          "temp_c": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Temp C"
          },
          "weight_kg": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          }
        },
        "required": [
          "id",
          "recorded_at",
          "bp_sys",
          "bp_dia",
          "pulse",
          "temp_c",
          "spo2",
          "rr",
          "weight_kg",
          "height_cm"
        ],
        "title": "VitalOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "HealthSaathi API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/break-glass": {
      "get": {
        "operationId": "break_glass_queue_api_v1_break_glass_get",
        "parameters": [
          {
            "in": "query",
            "name": "pending",
            "required": false,
            "schema": {
              "default": false,
              "title": "Pending",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BreakGlassItem"
                  },
                  "title": "Response Break Glass Queue Api V1 Break Glass Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Break Glass Queue",
        "tags": [
          "chart"
        ]
      }
    },
    "/api/v1/break-glass/{break_glass_id}/review": {
      "post": {
        "operationId": "review_break_glass_api_v1_break_glass__break_glass_id__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "break_glass_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Break Glass Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakGlassReviewIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BreakGlassItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Review Break Glass",
        "tags": [
          "chart"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/break-glass": {
      "post": {
        "operationId": "break_glass_api_v1_patients__patient_id__break_glass_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BreakGlassIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyChart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Break Glass",
        "tags": [
          "chart"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/chart": {
      "get": {
        "operationId": "chart_api_v1_patients__patient_id__chart_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "discriminator": {
                    "mapping": {
                      "admin": "#/components/schemas/AdminChart",
                      "doctor": "#/components/schemas/DoctorChart",
                      "lab": "#/components/schemas/LabChart",
                      "nurse": "#/components/schemas/NurseChart",
                      "patient": "#/components/schemas/PatientChart",
                      "reception": "#/components/schemas/ReceptionChart"
                    },
                    "propertyName": "view"
                  },
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ReceptionChart"
                    },
                    {
                      "$ref": "#/components/schemas/NurseChart"
                    },
                    {
                      "$ref": "#/components/schemas/LabChart"
                    },
                    {
                      "$ref": "#/components/schemas/DoctorChart"
                    },
                    {
                      "$ref": "#/components/schemas/AdminChart"
                    },
                    {
                      "$ref": "#/components/schemas/PatientChart"
                    }
                  ],
                  "title": "Response Chart Api V1 Patients  Patient Id  Chart Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "428": {
            "description": "A doctor must give a reason to open this chart."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chart",
        "tags": [
          "chart"
        ]
      },
      "post": {
        "operationId": "chart_with_reason_api_v1_patients__patient_id__chart_post",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccessReasonIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoctorChart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chart With Reason",
        "tags": [
          "chart"
        ]
      }
    },
...
    "/api/v1/patients/{patient_id}/emergency": {
      "get": {
        "operationId": "emergency_api_v1_patients__patient_id__emergency_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "break_glass_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Break Glass Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmergencyChart"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Emergency",
        "tags": [
          "chart"
        ]
      }
    },
...
  }
}